+++ src/core/configuration.cc
@@ -584,7 +584,7 @@ void Configuration::loadManagedClients(const JsonDoc& json, std::stringstream& e
             // no loggers array
             std::string defaultLogger = j.get<std::string>("default_logger", "");
             if (!defaultLogger.empty()) {
-                errorStream << "  Default logger ["  << defaultLogger << "] for client [" << clientId << "] is not part of [loggers] array. Please see https://github.com/amrayn/residue/blob/master/docs/CONFIGURATION.md#managed_clientsloggers";
+                errorStream << "  Default logger ["  << defaultLogger << "] for client [" << clientId << "] is not part of [loggers] array. Please see https://github.com/abumq/residue/blob/master/docs/CONFIGURATION.md#managed_clientsloggers";
             }
         }
     }