+++ src/net/http-client.cc
@@ -61,7 +61,7 @@ std::string HttpClient::fetchUrlContents(const std::string& url)
         curl_easy_setopt(curl, CURLOPT_URL, url.data());
         curl_easy_setopt(curl, CURLOPT_SSL_VERIFYPEER, 0L);
         curl_easy_setopt(curl, CURLOPT_SSL_VERIFYHOST, 0L);
-        curl_easy_setopt(curl, CURLOPT_USERAGENT, std::string("amrayn/residue/" +
+        curl_easy_setopt(curl, CURLOPT_USERAGENT, std::string("abumq/residue/" +
                                                               std::string(RESIDUE_VERSION_MAJOR) + "." +
                                                               std::string(RESIDUE_VERSION_MINOR) + "." +
                                                               std::string(RESIDUE_VERSION_PATCH)).data());