+++ tools/config-validator/main.cc
@@ -31,7 +31,7 @@ void displayUsage() {
 }
 
 void displayVersion() {
-    std::cout << "Config Validator v" << RESIDUE_VERSION << std::endl << std::endl << "Configuration file validator for residue logging server (https://github.com/amrayn/residue)" << std::endl;
+    std::cout << "Config Validator v" << RESIDUE_VERSION << std::endl << std::endl << "Configuration file validator for residue logging server (https://github.com/abumq/residue)" << std::endl;
 }
 
 int main(int argc, char* argv[])